requirements.txt
psycopg2==2.9.9
psycopg2_binary==2.9.9
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.38.0

